+++ src/JenkinsIndicator.ts
@@ -3,125 +3,153 @@
 *  Licensed under the MIT License. See License.md in the project root for license information.
 *--------------------------------------------------------------------------------------------*/
 
-import fs = require("fs");
-import path = require("path");
 import * as vscode from "vscode";
 import * as Jenkins from "./Jenkins";
+import { Setting } from "./setting";
 
 export class JenkinsIndicator {
 
-    private statusBarItem: vscode.StatusBarItem;
-    private currentStatus: Jenkins.JenkinsStatus = <Jenkins.JenkinsStatus> {};
-    private currentBasePath: string;
+    private statusBarItems: {[settingName: string]: vscode.StatusBarItem} = {};
+    private settingNameToUrl: {[settingName: string]: string} = {};
 
     public dispose() {
-        this.hideReadOnly();
+        this.hideReadOnly(this.statusBarItems);
     }
 
-    public updateJenkinsStatus(basePath: string) {
+    public updateJenkinsStatus(settings: Setting[], registerCommand: (cmd: string, callback: () => void ) => void, deRegisterCommand: (cmd: string) => void): Setting[] {        
+        let noNameCount = -1;
+        this.settingNameToUrl = {};
 
-        return new Promise((resolve, reject) => {
-            // Create as needed
-            if (!this.statusBarItem) {
-                this.statusBarItem = vscode.window.createStatusBarItem(vscode.StatusBarAlignment.Left);
-                this.statusBarItem.command = "jenkins.openInJenkinsConsoleOutput"
+        for (let index = 0; index < settings.length; index++) {
+            const setting = settings[index];
+            if (!(setting.name)) {
+                noNameCount++;
+                setting.name = "Jenkins " + (noNameCount ? noNameCount : "");
             }
 
-            // even if 'not available', it has to update the current 'updated'
-            this.currentBasePath = basePath;
+            this.settingNameToUrl[setting.name] = setting.url;
 
-            if (!fs.existsSync(path.join(basePath, ".jenkins"))) {
-                this.statusBarItem.tooltip = "No Jenkins defined for this project";
-                this.statusBarItem.text = "(No Jenkins)";
-                this.statusBarItem.show();
-                this.currentStatus = <Jenkins.JenkinsStatus> {};
-                resolve(true);
-                return;
+            // Create as needed
+            if (!this.statusBarItems[setting.name]) {
+                this.statusBarItems[setting.name] = vscode.window.createStatusBarItem(vscode.StatusBarAlignment.Left);
+                this.statusBarItems[setting.name].command = "Jenkins." + setting.name + ".openInJenkins";
+
+                registerCommand("Jenkins." + setting.name + ".openInJenkins", () => {
+                    vscode.env.openExternal(vscode.Uri.parse(this.settingNameToUrl[setting.name]));
+                });
+                registerCommand("Jenkins." + setting.name + ".openInJenkinsConsoleOutput", () => {
+                    jjj.getStatus(url, user, pw)
+                    .then((status) => {
+                        if (status.connectionStatus === Jenkins.ConnectionStatus.Connected) {
+                            vscode.env.openExternal(vscode.Uri.parse(this.settingNameToUrl[setting.name] + status.buildNr.toString() + "/console"));
+                        } else {
+                            vscode.window.showWarningMessage("The Jenkins job has some connection issues. Please check the status bar for more information.");     
+                        }   
+                    });
+                });
             }
 
             let jjj: Jenkins.Jenkins;
             jjj = new Jenkins.Jenkins();
 
-            let url: string;
-            let user: string;
-            let pw: string;
+            const url = setting.url;
+            const user = setting.username ? setting.username : "";
+            const pw = setting.password ? setting.password : "";
 
-            const settings = JSON.parse(fs.readFileSync(path.join(basePath, ".jenkins")).toString());
-            url = settings.url;
-            user = settings.username ? settings.username : "";
-            pw = settings.password ? settings.password : "";
-
-            if (settings.strictTls !== undefined) {
-                process.env.NODE_TLS_REJECT_UNAUTHORIZED = settings.strictTls ? "1" : "0";
+            if (setting.strictTls !== undefined) {
+                process.env.NODE_TLS_REJECT_UNAUTHORIZED = setting.strictTls ? "1" : "0";
             }
+
+            this.statusBarItems[setting.name].text = setting.name;
+            this.statusBarItems[setting.name].show();
 
             // invalid URL
             if (!url) {
-                this.statusBarItem.tooltip = "No URL Defined";
-                this.statusBarItem.text = "Jenkins $(x)";
-                this.statusBarItem.show();
-                this.currentStatus = <Jenkins.JenkinsStatus> {};
-                resolve(true);
-                return;
+                this.statusBarItems[setting.name].tooltip = "No URL Defined";
+                this.statusBarItems[setting.name].text = "Jenkins $(x)";
+                continue;
             }     
             
             jjj.getStatus(url, user, pw)
                 .then((status) => {
 
                     let icon: string;
-                    this.currentStatus = status;
+                    let tooltip = 
+                            "Job Name: " + status.jobName + "\n" +
+                            "Status: " + status.statusName + "\n" +
+                            "URL: " + status.url + "\n" +
+                            "Connection Status: " + status.connectionStatusName;
+                    
+                    if (status.buildNr !== undefined) {
+                        tooltip = tooltip + "\n" + 
+                        "Build #: " + status.buildNr;
+                    }
+
+                    if (status.code !== undefined) {
+                        tooltip = tooltip + "\n" + 
+                        "Code #: " + status.code;
+                    }
 
                     switch (status.status) {
-                        case Jenkins.BuildStatus.Sucess:
+                        case Jenkins.BuildStatus.InProgress:
+                            icon = " $(pulse)";
+                            break;
+
+                        case Jenkins.BuildStatus.Success:
                             icon = "$(check) ";
-                            this.statusBarItem.tooltip = 
-                                "Job Name: " + status.jobName + "\n" +
-                                "URL.....: " + status.url + "\n" +
-                                "Build #.: " + status.buildNr; 
                             break;
 
                         case Jenkins.BuildStatus.Failed:
                             icon = "$(alert) ";
-                            if (status.connectionStatus === Jenkins.ConnectionStatus.AuthenticationRequired) {
-                                this.statusBarItem.tooltip = 
-                                    "Job Name: " + status.jobName + "\n" +
-                                    "<<Authenthication Required>>"; 
-                            } else {
-                                this.statusBarItem.tooltip = 
-                                    "Job Name: " + status.jobName + " -- (FAILED)\n" +
-                                    "URL.....: " + status.url + "\n" +
-                                    "Build #.: " + status.buildNr;
-                            }
                             break;
                     
                         default:
                             icon = "$(stop) ";
-                            this.statusBarItem.tooltip = 
-                                "Job Name: " + status.jobName + "\n" +
-                                "URL.....: " + status.url + "\n" +
-                                "Build #.: " + status.buildNr; 
                     }
                         
-                    this.statusBarItem.text = icon + "Jenkins";
-                    this.statusBarItem.show();
-                    resolve(status !== undefined);
+                    this.statusBarItems[setting.name].text = icon + setting.name;
+                    this.statusBarItems[setting.name].tooltip = tooltip;
+                    this.statusBarItems[setting.name].show();
                 });
-        });
+        }
+
+        const tmpStatusBarItems = this.statusBarItems;
+        this.statusBarItems = {};
+        for (const key in this.settingNameToUrl) {
+            if (this.settingNameToUrl.hasOwnProperty(key)) {
+                this.statusBarItems[key] = tmpStatusBarItems[key];
+                delete tmpStatusBarItems[key];
+            }
+        }
+        
+        this.hideReadOnly(tmpStatusBarItems);
+        for (const key in tmpStatusBarItems) {
+            if (tmpStatusBarItems.hasOwnProperty(key)) {
+                deRegisterCommand("Jenkins." + key + ".openInJenkins");
+                deRegisterCommand("Jenkins." + key + ".openInJenkinsConsoleOutput");                
+            }
+        }
+
+        return settings;
     }
 
-    public hideReadOnly() {
-        if (this.statusBarItem) {
-            this.statusBarItem.dispose();
+    public hideReadOnly(items) {
+        for (const key in items) {
+            if (items.hasOwnProperty(key)) {
+                const statusBarItem = items[key];
+                statusBarItem.dispose();                
+            }
         }
     }
-    
-    public getCurrentStatus(): Jenkins.JenkinsStatus {
-        return this.currentStatus;
-    }
 
-    public getCurrentBasePath() {
-        return this.currentBasePath;
-    }
+    // public getCommandFromSettingName(settingName: string){
+    //     for (const key in this.settingNameToUrl) {
+    //         if (this.settingNameToUrl.hasOwnProperty(settingName)) {
+    //             return statusBarItemsthis.statusBarItems[key] = tmpStatusBarItems[key];
+    //             delete tmpStatusBarItems[key];
+    //         }
+    //     }
+    // }
 }
 
 export class JenkinsIndicatorController {